Can damage now

+++ Gladius/CommonWeapon/src/main/java/CommonWeapon/Weapon.java
@@ -2,11 +2,11 @@ package CommonWeapon;
 import Common.data.Entity;
 public abstract class Weapon extends Entity {
     private String name;
-    private float damage;
+    private int damage;
     private float weight;
     private final String texturePath;
 
-    public Weapon(String name, float damage, float weight, float range, String texturePath) {
+    public Weapon(String name, int damage, float weight, float range, String texturePath) {
         super(null, range, 9, 36, 0, 0.9f, 0.9f, 0);
         this.name = name;
         this.damage = damage;
@@ -18,7 +18,7 @@ public abstract class Weapon extends Entity {
         return this.name;
     }
 
-    public float getDamage() {
+    public int getDamage() {
         return this.damage;
     }
