+++ Gladius/WeaponComponent/src/main/java/Weapon/Spear.java
@@ -4,7 +4,7 @@ import CommonWeapon.Weapon;
 
 public class Spear extends Weapon {
 
-    protected Spear(String name, float damage, float weight, float range, String texturePath) {
+    protected Spear(String name, int damage, float weight, float range, String texturePath) {
         super(name, damage, weight, range, texturePath);
     }
