+++ Gladius/WeaponComponent/src/main/java/Weapon/Sword.java
@@ -5,7 +5,7 @@ import CommonWeapon.Weapon;
 public class Sword extends Weapon {
     private static final String texturePath = "swordMiniRect.png";
     private static Weapon sword = new Sword("Sword", 10, 8, 10);
-    private Sword(String name, float damage, float weight, float range) {
+    private Sword(String name, int damage, float weight, float range) {
         super(name, damage, weight, range, texturePath);
     }
