+++ Gladius/WeaponComponent/src/main/java/Weapon/Unarmed.java
@@ -4,7 +4,7 @@ import CommonWeapon.Weapon;
 
 public class Unarmed extends Weapon {
 
-    protected Unarmed(String name, float damage, float weight, float range, String texturePath) {
+    protected Unarmed(String name, int damage, float weight, float range, String texturePath) {
         super(name, damage, weight, range, texturePath);
     }
